requirements.txt
Pillow
pyyaml
rich
